fix: align receiver_id dtype with entity_id (int32, not int64)
entity_id is pa.int32() in the archetype schema and cast to int32
in the updater. receiver_id was parsed as int64, which could cause
join mismatches in production (silent delivery failures).

+++ src/archetype/app/messaging.py
@@ -147,7 +147,7 @@ class MessageDeliveryProcessor(AsyncProcessor):
         # --- 2. Parse JSON messages into structured fields ---
         msg_struct_type = DataType.struct(
             {
-                "receiver_id": DataType.int64(),
+                "receiver_id": DataType.int32(),
                 "channel": DataType.string(),
                 "content": DataType.string(),
                 "_append_history": DataType.bool(),
